requirements.txt
pandas
datetime
pytz
yfinance
openai
mysql-connector-python
pydantic
langchain
pymysql
scikit-learn
statsmodels
